Fix tests now that rich text bundle is asynchronously loaded
The fix was to just stub out the requirejs lazy load function and mock it so
it returns a rich text bundle stub.

For some reason, the tests don't work unless defining the functions in
cmEditor before the lazy load function. So I've wrapped in an init
function and called later. I'm not sure why this is needed to be honest