Hack: only show fuzzy matched list if the search has more than a backslash
Fuse seems to have a weird bug where if every item in the list matches equally
it picks an item further down the unsorted list, rather than the first of the
unsorted list. This appears like the completions are in a weird order when the
autocomplete is first opened.

To work around this, we check if a single backslash has been entered and show
only the unsorted list without running it through Fuse. If more has been
entered, we run through Fuse to get the fuzzy matching

Finally fuse is memoized to reduce calls to a somewhat expensive operation