Update variables in views for handler changes
All pages are now going through a handler, to properly handle metadata
and other shared functionality.

In `layout.pug`, the variable `metadata` will be picked up on the
parent layout.

About/General Pages - Data was moved out of `data`.
Blog - Remove unnecessary nesting in `fields`. This is how the API
returns data, but we can omit this for the layout to make logic clearer.

The preview UI check was updatd due to name and type change.

Update images alt tag on about page and add alt tag to images on
general pages. The `title` is required when uploading media the the
CMS, but `description` is not.