Make the admin_id subscription property non-editable on the admin panel
When we change the susbcription owner we must ensure the new admin is added
to the manager_ids array. The question of what to do with the previous admin
is unclear, we could keep it as team manager or remove it altogether.

We still need to develop the CRUD UI to add/remove managers from a team.
When we do that, we'll have a better answer to these questions.

In the meantime, I've made the field non-editable, similar to group
members.