Change teams representation in segmentation responses
Web knows about the mapping between v1 and v2 ids for teams, so
it's easier if we include that information in the segemntation
responses.

Before we had a response like:

{
  id : 1,
  teamIds: [3, 4],
  affiliationIds: [5, 7],
  v2TeamIds: ['5ad60490c34621025f26006a', '5ae8366062616006e8eb07d0']
}

Which is arguably confusing, given the asymmetry between teamIds (which
are v1 ids) and v2TeamIds.

The new representation is:

{
  id : 1,
  teamIds: [3, 4],
  affiliationIds: [5, 7],
  teams: [
    { v1Id: 3, v2Id: '5ad60490c34621025f26006a'},
    { v2Id: '5ae8366062616006e8eb07d0'},
    { v1Id: 4 }
  ]
}

The teamIds key is kept for compatibility but the plan is to remove it
when analytics is changed to rely only in the 'teams' property.