User record always updated when v1 user is created
Previously we would throw an error if the user was not created. This
code path was called whether or not we thought the account was linked.
This meant that when handling a conflict (i.e. the user had already
linked) it would throw an error.

We decided it was safer to touch the DB record even if there was a
conflict.