Set HTTP headers only once for portals redirects
When the URL path (`/edu/slug` or `/org/slug`) does not match the correct portal type
the user is redirected via a 301 HTTP request.

The HTTP headers were getting set twice when this happened. To resolve, the request to render the layout were moved to within an else statement for when the URL path is correct