Move route to skip CSRF but still get sessions
Fixes an issue where POSTing to /sign_in_to_v1 would not pick up the
users session. This meant that users wouldn't actually be signed into v1

However we still need to disable CSRF protection as some redirects are
used by third parties (and thus don't have CSRF tokens). The CSRF
protection is done on the v1 side for the only redirect that currently
needs this.