Localise and rebrand bonus page links
Existing links had hardcoded English text and mentioned Sharelatex instead of Overleaf. This wraps the content in `translate()` calls so that the text can be localised and re-branded appropriately.

Removed G+ support because it's going away and nobody uses it.

Changed the Facebook app ID as the URL is now Overleaf and the callback
needs to be different.

bug: overleaf/sharelatex#992