Allow project collaborators with write access to submit/export
Previously user had to be an owner or project admin to successfully submit, though the Submit option was offered to them even if they were only a collaborator with write access.

This change should make the experience consistent: if you have project write access not only will you be shown the Submit option, but it will work.