Add several helper classes and methods to assist sharing of functionality
The "open in overleaf" API wants to call into a number of things that are currently in more protected places, or duplicate code that already exists in other methods. This moves these out into helper classes to make this more transparent.

Also refactored ProjectHelper to add a createProjectFromSnippet method and reduce code duplication

bug: sharelatex/web-sharelatex-internal#731