Add "open-in-overleaf" module, part 1
This implements the basic functionality of the "open in overleaf" API (via the '/docs' endpoint.)

Currently you can post .tex snippets and URLs to '/docs' and have them open as separate projects. There is a middleware layer that handles forwarding requests via login or with csrf tokens. This does not yet include the full functionality of the v1 API, and so should still be hidden behind a redirect for now.

bug: sharelatex/web-sharelatex-internal#731