Redir to ol_doc_id when import completed
Previously we would always try to redirect to the project mongo id, as
this always does the "right thing" based on the user's permissions.

However, when importing via a read-only link, the user's access is never
recorded on the v1 side and therefore can't be imported to v2.

We decided against the alternative of adding the importing user to the
project's tokenAccessReadOnly_refs, because this would require special
handling for read-only links only.