Wrap request in Promise to force async behaviour
In our version of jQuery, $.Deferred() does not fully mimic the Promise
spec, and specifically it does not force async resolution of
Promise.resolve(). This meant that in the tests, both requests would
resolve synchronously. Then because React batches this.setState(), the
initiated export state would be skipped.

By wrapping the request in a native Promise, we force async behaviour
in request resolution.