Fix BrandedPublishModal showing download prompt in guide html
publish_guide_html can have a DOWNLOAD placeholder that is supposed to
be replaced with a download zip button. This was handled correctly in
the GuidePublishModal, but not the BrandedPublishModal. This PR fixes
by reusing the same PublishGuide component that is used in the
GuidePublishModal to show the download button.

This required a v1 change to send the default brand variation id for the
brand_data object.