Use constant-time comparison to find project by read-and-write token
This means matching on the numeric prefix of the token, then doing a
constant-time string comparison on the entire token. This is a little more
awkward than just matching straight on the whole token, but is more secure.

This change will require a migration to pluck out the numeric prefix for
all projects that have read-write tokens