Exports create API returns error status on error
Return an error status from the export creation endpoint when v1 returns an error status to us.

Previously when v1 responded with an error, e.g. status `422` body `{"status":422,"message":"This project has already been submitted to this journal"}`, we returned the same message body to the publish modal, but returned `200` as the status of the v2 API call. Now we would return `422` in this situation instead.