Detect encoding of non-utf8 files and convert during load
Uses the 'is-utf8' package to detect whether the file is utf8. If not,
it assumes utf16 if there is a unicode BOM, otherwise loads as latin1.

Also added .lbx, .bbx, .cbx and .m to the list of text file types that
we handle, and added support for files with \r line endings.

bug: overleaf/sharelatex#1413
bug: overleaf/sharelatex#1566
bug: overleaf/sharelatex#889